pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.24.0",
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
